Consumable and Consumer work (No animation just right click on inspector and consume) Fixed health script
-hunger bar
-Consumable script works in Ben's scene just right click in Unity inspector and hit consume.

+++ Assets/healthBar.cs
@@ -14,7 +14,7 @@ public class healthBar : MonoBehaviour
 
         fill.color = gradient.Evaluate(1f); //max health at 1, 0 for min health 
     }
-    public void SetHealth (int health) 
+    public void SetHealth (float health) 
     {
         slider.value = health; 
         fill.color = gradient.Evaluate(slider.normalizedValue);
